接口封装
