+++ 20188376/src/ComprehensiveHeadFile.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <iostream>
+#include <fstream>
+#include <string>
+#include <iterator>
+#include <sstream>
+#include <vector>
+#include <algorithm>
+#include <map>
+
+using namespace std;
